+++ packages/ui/src/components/accordion/components/accordion-item/accordion-item.styles.ts
@@ -7,7 +7,8 @@ export const styles = tv(
       itemHeader: 'typography-body-9 flex w-full flex-1 items-center justify-between px-3 py-2 group-first:border-t-0',
       headerTitleWrapper: 'flex-1 pr-2 text-left',
       indicator: 'size-3 rotate-90',
-      content: '',
+      content: 'bg-white [&_:focus-visible]:focus-outline',
+      motionContent: '',
     },
     variants: {
       look: {
@@ -30,6 +31,14 @@ export const styles = tv(
         },
         false: {
           content: 'hidden',
+        },
+      },
+      isClosed: {
+        true: {
+          motionContent: 'overflow-hidden',
+        },
+        false: {
+          motionContent: 'overflow-visible',
         },
       },
       isDisabled: {
@@ -69,6 +78,12 @@ export const styles = tv(
         isOpen: true,
         rounded: true,
         className: 'group-last:rounded-none',
+      },
+      {
+        slots: ['content'],
+        isOpen: true,
+        rounded: true,
+        className: 'group-last:rounded-b',
       },
       { slots: ['content'], isOpen: true, look: 'lego', className: 'mt-[1px] border-l-[0.375rem] border-r' },
     ],