+++ packages/ui/src/components/button-dropdown/button-dropdown.types.ts
@@ -1,7 +1,9 @@
 import { ButtonHTMLAttributes, ReactNode } from 'react';
+import { Placement } from 'react-aria';
 import { type VariantProps } from 'tailwind-variants';
 
 import { ButtonProps } from '../button/index.js';
+import { IconProps } from '../icon/icon.types.js';
 
 import { styles } from './button-dropdown.styles.js';
 import { ButtonDropdownPanelProps } from './components/button-dropdown-panel/button-dropdown-panel.types.js';
@@ -20,11 +22,15 @@ export type ButtonDropdownProps = {
   /**
    * Look for button
    */
-  look?: 'primary' | 'hero' | 'faint';
+  look?: 'primary' | 'hero' | 'faint' | 'unstyled' | 'link';
   /**
    * State of whether the Popover is open
    */
   open?: boolean;
+  /**
+   * placement of the popover
+   */
+  placement?: Placement;
   /**
    * Soft look button
    */
@@ -32,7 +38,11 @@ export type ButtonDropdownProps = {
   /**
    * Button text
    */
-  text: string;
+  text?: ReactNode;
+  /**
+   * Dropdown icon
+   */
+  dropDownIcon?: (props: IconProps) => ReactNode;
 } & ButtonHTMLAttributes<Element> &
   Pick<ButtonProps, 'size' | 'iconBefore' | 'block'> &
   Pick<ButtonDropdownPanelProps, 'shouldFlip'>;