+++ packages/ui/src/components/button/button.styles.ts
@@ -53,6 +53,9 @@ export const styles = tv(
         true: { base: 'focus-outline' },
         false: { base: 'outline-none' },
       },
+      removeLinkPadding: {
+        true: '',
+      },
     },
     compoundSlots: [
       {
@@ -72,6 +75,12 @@ export const styles = tv(
         look: 'faint',
         soft: true,
         className: 'hover:bg-light active:bg-light',
+      },
+      {
+        slots: ['base'],
+        look: 'link',
+        removeLinkPadding: true,
+        className: 'px-0',
       },
       {
         slots: ['iconBefore'],