+++ packages/ui/src/components/header/header.stories.tsx
@@ -26,8 +26,6 @@ type Story = StoryObj<typeof meta>;
  * > Default usage example
  */
 export const DefaultStory: Story = {
-  // for underscored property
-
   render: ({ brand: _, ...rest }, { globals: { theme } }) => {
     const brand = theme ? theme.toLowerCase() : 'wbc';
     return (
@@ -39,11 +37,23 @@ export const DefaultStory: Story = {
 };
 
 /**
+ * > Example of header with no logo link
+ */
+export const NoLogoLink: Story = {
+  render: ({ brand: _, ...rest }, { globals: { theme } }) => {
+    const brand = theme ? theme.toLowerCase() : 'wbc';
+    return (
+      <>
+        <Header disableLogoLink brand={brand === 'btfg' ? 'wbc' : brand} {...rest} />
+      </>
+    );
+  },
+};
+
+/**
  * > Example with a button on the right
  */
 export const RightButton: Story = {
-  // for underscored property
-
   render: ({ brand: _, ...rest }, { globals: { theme } }) => {
     const brand = theme ? theme.toLowerCase() : 'wbc';
     return (
@@ -62,8 +72,6 @@ export const RightButton: Story = {
  * > Example of logo centering when XSL
  */
 export const CenterAtXS: Story = {
-  // for underscored property
-
   render: ({ brand: _, ...rest }, { globals: { theme } }) => {
     const brand = theme ? theme.toLowerCase() : 'wbc';
     return (
@@ -78,8 +86,6 @@ export const CenterAtXS: Story = {
  * > Example of logo with onClick
  */
 export const LogoOnClick: Story = {
-  // for underscored property
-
   render: ({ brand: _, ...rest }, { globals: { theme } }) => {
     const brand = theme ? theme.toLowerCase() : 'wbc';
     return (
@@ -95,8 +101,6 @@ export const LogoOnClick: Story = {
  * > Example of header with skiplink
  */
 export const LogoWithSkipLink: Story = {
-  // for underscored property
-
   render: ({ brand: _, ...rest }, { globals: { theme } }) => {
     const brand = theme ? theme.toLowerCase() : 'wbc';
     return (
@@ -111,8 +115,6 @@ export const LogoWithSkipLink: Story = {
  * > Example fixed header. Does not show correctly in Docs view as it will show how it looks when scrolled, check specifc story for non-scrolled view
  */
 export const Fixed: Story = {
-  // for underscored property
-
   render: ({ brand: _, ...rest }, { globals: { theme } }) => {
     const brand = theme ? theme.toLowerCase() : 'wbc';
     return (
@@ -127,8 +129,6 @@ export const Fixed: Story = {
  * > Example of header with arrow button
  */
 export const WithArrow: Story = {
-  // for underscored property
-
   render: ({ brand: _, ...rest }, { globals: { theme } }) => {
     const brand = theme ? theme.toLowerCase() : 'wbc';
     return (
@@ -143,8 +143,6 @@ export const WithArrow: Story = {
  * > Example of header with arrow button with onClick
  */
 export const WithArrowOnClick: Story = {
-  // for underscored property
-
   render: ({ brand: _, ...rest }, { globals: { theme } }) => {
     const brand = theme ? theme.toLowerCase() : 'wbc';
     return (
@@ -165,13 +163,12 @@ export const WithArrowOnClick: Story = {
  * > Example of header with hamburger (only visible below xsl)
  */
 export const WithHamburger: Story = {
-  // for underscored property
-
   render: ({ brand: _, ...rest }, { globals: { theme } }) => {
     const brand = theme ? theme.toLowerCase() : 'wbc';
     return (
       <>
         <Header leftIcon="hamburger" brand={brand === 'btfg' ? 'wbc' : brand} {...rest} />
+        <p>NOTE: Hamburger button only shows in sm breakpoint or smaller</p>
       </>
     );
   },
@@ -181,8 +178,6 @@ export const WithHamburger: Story = {
  * > Example of header with hamburger (only visible below xsl) with onClick
  */
 export const WithHamburgerOnClick: Story = {
-  // for underscored property
-
   render: ({ brand: _, ...rest }, { globals: { theme } }) => {
     const brand = theme ? theme.toLowerCase() : 'wbc';
     return (
@@ -194,6 +189,7 @@ export const WithHamburgerOnClick: Story = {
           brand={brand === 'btfg' ? 'wbc' : brand}
           {...rest}
         />
+        <p>NOTE: Hamburger button only shows in sm breakpoint or smaller</p>
       </>
     );
   },