+++ packages/ui/src/components/header/header.types.ts
@@ -13,6 +13,11 @@ export type HeaderProps = {
    */
   brand: Exclude<BrandKey, 'btfg'>;
   /**
+   * Removes anchor link from logo
+   * @default false
+   */
+  disableLogoLink?: boolean;
+  /**
    * Enable fixed header
    */
   fixed?: boolean;
@@ -26,7 +31,8 @@ export type HeaderProps = {
    */
   isScrolled?: boolean;
   /**
-   * Visually hidden text for left button
+   * Aria-label for the arrow/hamburger button
+   * @default leftIcon === 'arrow' ? 'Back' : 'Menu'
    */
   leftAssistiveText?: string;
   /**
@@ -46,7 +52,8 @@ export type HeaderProps = {
    */
   logoCenter?: boolean;
   /**
-   * Link for logo
+   * Link for logo, can be disabled with disableLogoLink
+   * @default '#'
    */
   logoLink?: string;
   /**