+++ packages/ui/src/components/modal/components/modal-dialog/components/modal-dialog-body/modal-dialog-body.styles.ts
@@ -2,16 +2,58 @@ import { tv } from 'tailwind-variants';
 
 export const styles = tv(
   {
-    slots: { base: 'flex-1 overflow-auto' },
+    slots: { base: 'flex-1 transition-shadow delay-0 duration-200 ease-[ease]' },
     variants: {
       size: {
         full: { base: 'px-4 py-3' },
         lg: { base: 'px-12 pb-12' },
         md: { base: 'px-7 pb-7' },
-        sm: { base: 'px-5 pb-7' },
-        fluid: { base: 'px-5 pb-7' },
+        sm: { base: 'px-5 pb-5' },
+        fluid: { base: 'px-5 pb-5' },
+      },
+      canScroll: {
+        true: {
+          base: 'shadow-[0px_-4px_5px_-2px_inset] shadow-black/30',
+        },
+      },
+      scrollAtBottom: {
+        true: { base: 'shadow-[0_0_0_0_inset]' },
+      },
+      compact: {
+        true: '',
+        false: '',
+      },
+      footerPresent: {
+        true: '',
+        false: '',
       },
     },
+    compoundSlots: [
+      {
+        slots: ['base'],
+        size: ['lg'],
+        compact: true,
+        className: 'overflow-y-auto px-5 pb-3',
+      },
+      {
+        slots: ['base'],
+        size: ['md'],
+        compact: true,
+        className: 'overflow-y-auto px-5 pb-2',
+      },
+      {
+        slots: ['base'],
+        size: ['md'],
+        footerPresent: true,
+        className: 'pb-5',
+      },
+      {
+        slots: ['base'],
+        size: ['lg'],
+        footerPresent: true,
+        className: 'pb-6',
+      },
+    ],
   },
   { responsiveVariants: ['xsl', 'sm', 'md', 'lg', 'xl'] },
 );