+++ packages/ui/src/components/modal/components/modal-dialog/modal-dialog.styles.ts
@@ -3,14 +3,14 @@ import { tv } from 'tailwind-variants';
 export const styles = tv(
   {
     slots: {
-      base: 'relative mx-auto flex max-w-full flex-col overflow-hidden rounded bg-white text-text outline-none',
-      title: 'typography-body-7 pb-4 pt-9 font-bold text-text',
+      base: 'relative flex max-h-full max-w-full flex-col rounded bg-white text-text outline-none',
+      title: 'typography-body-7 pb-4 pt-9 font-bold text-text transition-shadow delay-0 duration-200 ease-[ease]',
       close: 'absolute right-0 top-0 block p-3',
     },
     variants: {
       size: {
         full: {
-          base: 'max-h-screen w-full flex-1',
+          base: 'h-screen max-h-screen w-full flex-1 rounded-none',
           close: 'p-2',
           title: 'px-4 py-3',
         },
@@ -20,22 +20,45 @@ export const styles = tv(
         },
         md: {
           base: 'w-[37.5rem]',
-          title: 'px-7 ',
+          title: 'px-7',
         },
         sm: {
           base: 'w-[25rem]',
           title: 'px-5',
         },
         fluid: {
-          base: 'w-full',
+          base: 'w-screen',
           title: 'px-5',
         },
       },
       isFocusVisible: {
-        true: { close: 'focus-outline' },
+        true: { close: '!outline-offset-[-3px] focus-outline' },
         false: { close: 'outline-none' },
       },
+      compact: {
+        true: '',
+        false: '',
+      },
+      scrolled: {
+        true: {
+          title: 'shadow-[0px_2px_5px_0px] shadow-black/30',
+        },
+      },
     },
+    compoundSlots: [
+      {
+        slots: ['base'],
+        size: ['md', 'lg'],
+        compact: true,
+        className: 'max-h-[80vh] overflow-hidden',
+      },
+      {
+        slots: ['title'],
+        size: ['lg', 'md'],
+        compact: true,
+        className: 'min-h-[90px] px-5 pb-4 pt-6',
+      },
+    ],
   },
   { responsiveVariants: ['xsl', 'sm', 'md', 'lg', 'xl'] },
 );