+++ packages/ui/src/components/tabs/components/tabs-tab-panel/tabs-tab-panel.styles.ts
@@ -2,7 +2,7 @@ import { tv } from 'tailwind-variants';
 
 export const styles = tv(
   {
-    base: 'flex-1 bg-white p-4',
+    base: 'flex-1 bg-white p-4 [&_:focus-visible]:focus-outline',
     variants: {
       look: {
         default: 'border border-border',