+++ packages/ui/src/components/tooltip/components/tooltip-content/tooltip-content.component.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import { styles } from './tooltip-content.styles.js';
+import { TooltipContentProps } from './tooltip-content.types.js';
+
+export function TooltipContent({ children, id }: TooltipContentProps) {
+  return children ? (
+    <span className={styles()} role="tooltip" id={id}>
+      {children}
+    </span>
+  ) : null;
+}